pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpang-jehyu-mvp"
version = "0.1.0"
description = "Affiliate marketing automation MVP for Threads + Instagram"
requires-python = ">=3.11"
dependencies = [
  "alembic==1.14.1",
  "cryptography==44.0.1",
  "fastapi==0.115.8",
  "google-cloud-tasks==2.18.0",
  "google-cloud-storage==2.19.0",
  "httpx==0.28.1",
  "itsdangerous==2.2.0",
  "jinja2==3.1.6",
  "pillow==11.1.0",
  "psycopg[binary]==3.2.6",
  "pydantic-settings==2.8.1",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.38",
  "uvicorn[standard]==0.34.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
